requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
slowapi==0.1.8
python-multipart==0.0.6
email-validator==2.1.0.post1
httpx==0.25.2
onnxruntime-gpu==1.16.3
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
celery==5.3.6
redis==5.0.1
trimesh==4.0.5
scikit-image==0.22.0
pyrender==0.1.45
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
playwright==1.40.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
regex==2023.10.3
